Upgrade dependencies, fix tests, use through2
prova is not installable anymore. it depends on ansi-codes that
has been removed from npm.
Replaced it with tape.

Using through2 instead of through to support pull streams
rather than push streams.

The tests were ending on each sampled piece of data.
It was not a problem with through but executing with through2
things became more asynchroneous and was failing the test.
Also the second test was sampling only once. Now we do sample
as many as the test intended.